agents/content/requirements.txt
# OpenClaw Content Department — Python dependencies
# Install with: pip install -r requirements.txt

# ── Core framework ─────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Database ───────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1

# ── Async task queue ───────────────────────────────────────────────────────
celery[redis]>=5.4.0
redis>=5.0.4

# ── Vector DB ──────────────────────────────────────────────────────────────
qdrant-client>=1.9.1

# ── AI / ML ────────────────────────────────────────────────────────────────
# Google Gemini — content writing and briefing
google-generativeai>=0.8.0

# OpenAI — embeddings (text-embedding-3-small) + DALL-E 3
openai>=1.30.0

# ── Trend APIs ─────────────────────────────────────────────────────────────
praw>=7.7.1              # Reddit API
pytrends>=4.9.2          # Google Trends (unofficial)
httpx>=0.27.0            # Async HTTP for Ahrefs / GSC REST

# ── Media / CMS ────────────────────────────────────────────────────────────
cloudinary>=1.40.0

# ── NLP / SEO utilities ────────────────────────────────────────────────────
beautifulsoup4>=4.12.3
lxml>=5.2.2
textstat>=0.7.3

# ── Real-time (FastAPI WebSocket / Socket.IO) ──────────────────────────────
python-socketio>=5.11.2
python-multipart>=0.0.9

# ── Observability ──────────────────────────────────────────────────────────
structlog>=24.2.0

# ── Dev / Testing ──────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0
python-dotenv>=1.0.1
